Implement the ability to notify changes to the web container thread pool under the Config mode. (#1133)
* Add notification logic after modifying web thread pool parameters.

* Refactor the notification module to enable notification capability for the web container.

* Add application profile enum.

* Revert the bugfix changes and split them into issue #1134.

* Revert @Order annotation in AbstractWebThreadPoolService.

* fix profile enum name with test environment.

* Fix: remove unused imports.

* Delete ProfileEnum class.

* Modify the way of obtaining the ID of the web thread pool.

* Move the IExecutorProperties class to the common module.

* Narrow the scope of @SuppressWarnings annotation usage.

+++ hippo4j-adapter/hippo4j-adapter-web/src/main/java/cn/hippo4j/adapter/web/undertow/UndertowWebThreadPoolHandlerSupport.java
@@ -82,7 +82,7 @@ public class UndertowWebThreadPoolHandlerSupport implements IWebThreadPoolHandle
             XnioWorker xnioWorker = (XnioWorker) executor;
             int minThreads = xnioWorker.getOption(Options.WORKER_TASK_CORE_THREADS);
             int maxThreads = xnioWorker.getOption(Options.WORKER_TASK_MAX_THREADS);
-            int keepAliveTime = xnioWorker.getOption(Options.WORKER_TASK_KEEPALIVE);
+            long keepAliveTime = xnioWorker.getOption(Options.WORKER_TASK_KEEPALIVE);
             parameterInfo.setCoreSize(minThreads);
             parameterInfo.setMaxSize(maxThreads);
             parameterInfo.setKeepAliveTime(keepAliveTime);
@@ -135,7 +135,7 @@ public class UndertowWebThreadPoolHandlerSupport implements IWebThreadPoolHandle
             XnioWorker xnioWorker = (XnioWorker) executor;
             Integer coreSize = threadPoolParameterInfo.corePoolSizeAdapt();
             Integer maxSize = threadPoolParameterInfo.maximumPoolSizeAdapt();
-            Integer keepAliveTime = threadPoolParameterInfo.getKeepAliveTime();
+            int keepAliveTime = threadPoolParameterInfo.getKeepAliveTime().intValue();
             int originalCoreSize = xnioWorker.getOption(Options.WORKER_TASK_CORE_THREADS);
             int originalMaximumPoolSize = xnioWorker.getOption(Options.WORKER_TASK_MAX_THREADS);
             int originalKeepAliveTime = xnioWorker.getOption(Options.WORKER_TASK_KEEPALIVE);
