+++ hippo4j-common/src/main/java/cn/hippo4j/common/model/ThreadPoolParameter.java
@@ -76,7 +76,7 @@ public interface ThreadPoolParameter {
      *
      * @return
      */
-    Integer getKeepAliveTime();
+    Long getKeepAliveTime();
 
     /**
      * Get execute time out
