+++ hippo4j-server/hippo4j-config/src/main/java/cn/hippo4j/config/model/ConfigInfoBase.java
@@ -81,7 +81,7 @@ public class ConfigInfoBase implements Serializable {
     /**
      * keepAliveTime
      */
-    private Integer keepAliveTime;
+    private Long keepAliveTime;
 
     /**
      * executeTimeOut
